move history sync tests into their own file
